ZERO-161 [Rename]: UI 폴더 세분화 및 파일 이동
