+++ src/components/tasks/Layout/TaskDate.tsx
@@ -1,8 +1,8 @@
 import { formatTaskListDate } from '@utils/getFormattedDate';
 import { useDate } from '@/src/contexts/DateContext';
-import Pagination from '../Pagination';
-import CalenderButton from '../UI/CalenderButton';
-import AddTaskListButton from '../UI/AddTaskListButton';
+import Pagination from '../UI/Pagination';
+import CalenderButton from '../UI/button/CalenderButton';
+import AddTaskListButton from '../UI/button/AddTaskListButton';
 
 export default function TaskDate() {
   const { date } = useDate();
